booking_assistant_project/requirements.txt
streamlit
google-generativeai
python-dotenv
sqlalchemy
pydantic
PyPDF2
email-validator
scikit-learn
tqdm
requests
